requirements.txt
python-docx>=0.8.11
pdfplumber>=0.7.6
requests>=2.31.0
pytest>=7.0.0
typing-extensions>=4.5.0
dotenv
gigachat